refactor(pause): introduce PauseController as drop-in for PauseManager
Phase 1 of pause-flow refactor (see plan in PR description). Zero
behavior change — pure rename + module shim so the multi-phase
refactor can build on PauseController incrementally without breaking
existing callers in one go.

- services/pause_controller.py: new module hosting PauseController
  (= old PauseManager logic verbatim) and singleton pause_controller.
- services/pause_manager.py: thin shim re-exporting the controller
  under the old names; existing imports
  ``from services.pause_manager import pause_manager`` keep working.
- tests/test_services/test_pause_controller.py: renamed from
  test_pause_manager.py via git mv; import + docstring updated to
  reference the new module.

Verify: 950/950 tests pass (full test_services + test_routes suite).
